Make the GUIDs in various data structures e_guid_t's rather than arrays of 16 bytes.  Use "sizeof" for the size of e_guid_t's, and use structure assignment to copy GUID values.
Make functions such as append_h225ras_call() and new_h225ras_call() take
pointers to e_guid_t's as arguments.

Define GUID_LEN in epan/guid-utils.h and use it as the length of a GUID
in a packet.  (Note that "sizeof e_guid_t" is not guaranteed to be 16,
although it is guaranteed to be the size of an e_guid_t.)

When constructing a display filter that matches a GUID, use
guid_to_str() to construct the string for the GUID.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17676 f5534014-38df-0310-8fa8-9805f1628bb7