change TRY_SCSI_SHORT_TRANSFER   to use the allocation length specified in the CDB instead of the allocation length from the data pdu   since sometimes the datapdu will be completely empty and thus we cant extract it from the data pdu. (unclear if this is really allowed by scsi   but anyway) this should make the scsi transfer length checking more correct.
update mmc  getconfiguration   and spc  inquiry  to use these macros
and flag these packets that are short due to a too small alloc_len
as truncated due to scsi  and not being short/malformed packets.


at the same time,  get rid of a few horrible   offset+x   in the inquiry dissector and replace with proper offset incrementation instead.






git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17784 f5534014-38df-0310-8fa8-9805f1628bb7