From Martin Mathieson: This patch fixes the decode problem mentioned in this bug report:
"3. Frames 9 and 13 are only partially decoded and show up as "Packet size limited during capture" It looks like the decode chokes on the DQ-RI line."

DQ-RI is a localconnectionoptions parameter and shouldn't appear as a top-level parameter. This patch makes sure that all of the single-character top-level parmeter codes are followed immediately by a ':'.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17816 f5534014-38df-0310-8fa8-9805f1628bb7