change iscsi to track transaction data in a fc_exchange_data structure   just as the fc / fcp transport does.
while reusing this structure for multi-purpose would make it berret suited to move the struct from packet-fc.h to packet-scsi.h
we would have to include packet-scsi.h   for everyone that includes packet-fc.h  and that are a few dependencies.
(would have to include it even in packet-llc.c)

so leave it as a fc structure for now.





git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17833 f5534014-38df-0310-8fa8-9805f1628bb7