Use 2-space indentation.
Clean up the checks for STP vs. RSTP vs. MSTP.

Show the version 3 length field as a separate field, rather than as the
top-level item for the MSTP stuff.

"Trust" the version 3 length field, so that if it doesn't agree with the
packet length, we report a malformed frame (as we should).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17973 f5534014-38df-0310-8fa8-9805f1628bb7