From Martin Mathieson:
This patch should hopefully remove any possible buffer overflows in 
parse_line() as reported by the current Coverity scan.  I'm not sure 
that the error it currently reports is valid (I think its confused by 
supposing that a condition that is being tested can be true, whereas it 
can't...), but this patch fixes a number of potential problems remaining 
in the function.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@17979 f5534014-38df-0310-8fa8-9805f1628bb7