Clean up indentation; use 2-space indentation throughout.
Save the media encoding name in the transport_info_t structure rather
than in a global variable.  Allocate it with tvb_get_ephemeral_string()
so it's released after the packet is completely processed.  Do *NOT*
assume it's necessarily non-null in decode_sdp_fmtp(), as the code flow
doesn't guarantee that to be true.

proto_tree_add_string() now suffices for adding a particular SDP item -
strings are now displayed "safely", with escaping of non-printable
characters done.

Update a comment (we no longer have the Big Transfer Vector to allow
plugins to call dissector functions in Windows, we have those functions
in a DLL).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18031 f5534014-38df-0310-8fa8-9805f1628bb7