force the AssertionValue type to be displayed as a STRING instead of as hex BYTES.
Most of the time  AssertionValue will contain an ascii string   so make it always display as a string to make the display "correct" most of the time insterad of being "wrong" most of the time.

There are situations when AssertionValue contains binary data though  and in those cases the display will be "wrong" (but not more wrong than the old dissector anyway)


What someone really should do (someone interested in ldap that is) to make it more correct would be to implement a dissector for AssertionValue in the template file and having the dissector check if any of the bytes of the octet string has a value <32   and if so display it in hex as   0x.....
It all bytes have values >=32   then it shoudlk display it as a string "...
instead.

Someone interested in ldap may spend time on this refinement.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18089 f5534014-38df-0310-8fa8-9805f1628bb7