initial bluetooth rfcomm dissector.
this dissector will not yet detect when ppp is passed over the rfcomm link
but the old code to detect and deescapt the ppp data is still in the dissector, though ifdeffed out   to serve as inspiration when ppp over rfcomm captures are made available.


the only captures i have with rfcomm are for raw serial communications so they dont contain any ppp frames. :-(




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18221 f5534014-38df-0310-8fa8-9805f1628bb7