From Alejandro Vaquero:
The "listen_rtp" plugin that allows to listen audio RTP 
conversations.

    The plugin is integrated to the "Voip Calls" feature. There is a new 
"listen" button in the "Voip Calls" that once the calls are selected and 
the "listen" is clicked, a new window will open. In this window you can 
change the simulated jitter buffer to be used for decoding the RTP 
packets. In this first implementation, only a static jitter buffer can 
be simulated. Then pressing "decode" will decode all the RTP and 
generate the graphical view of the audio channels. From there you can 
select up to two channels at the same time (to be played in the left and 
right channels) and then play, pause, stop,etc.....ok, the attached 
screenshot it is self explanatory (hopefully).
    The only codecs available from now are G711u and G711a law.

    The PortAudio (www.portaudio.com) library is used to play audio. 
This is an open source cross-platform Audio library. I have tested this 
on Windows XP  and in a linux FC4.

NOTE: only the plugin soure files checked in at this time.

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18588 f5534014-38df-0310-8fa8-9805f1628bb7