fropm martin m
fix for h450 to prevent an assertion for uninitialized hffields



Thanks for the capture, Keith.  The problem was with h450 hf fields that
weren't initialised sucessfully (at all in one case, or with non-unique
filter strings in several others) - it was hitting an assertion in proto.c
when an attempt was made to use those fields.

I was able to test by editing packet-h450.c directly, I couldn't regenerate
it from packet-h450-template.c.  I'm attaching a patch to
packet-h450-template.c that hopefully does the same thing.  If someone can
generate and check it packet-h450.c in for me I'll retest.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18804 f5534014-38df-0310-8fa8-9805f1628bb7