testing the win32 decryp[tion out  i noticed a lot of new previously not decoded info in the captures for constrained delegation on the Kerberos wiki page.
a new bit 0x00020000 is usde in the TGS-REQ packets and this results in a return of a PAC containing an unknown type 11 field.


the blob in the pac is 200 bytes and NDR encoded.   its structure is obvious since it contains 2 conformant and varying arrays  and three unique pointers.



enable decoding of this new KDCOptions bit and call it "constrained delegation"





git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18857 f5534014-38df-0310-8fa8-9805f1628bb7