Stephen Fisher <stephentfisher@yahoo.com>
Attached is a patch to packet-http.c that calls a subdissector for
traffic flowing through a proxy via the HTTP CONNECT method.  Most
protocols, especially SSL, can be tunneled through an HTTP proxy.
Wireshark currently says this traffic is "Continuation or non-HTTP
traffic" but this patch turns the payload over to the dissector for the
protocol being tunneled.  This is similar to how the Socks dissector
works.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18901 f5534014-38df-0310-8fa8-9805f1628bb7