Mark Lewis <mlewis@altera.com>
I have developed a plugin for Pro-MPEG FEC packets over RTP (see
previous posts on ethereal-dev). I have added a page and example capture
file to the Wiki (http://wiki.wireshark.org/2dParityFEC). The source and
Windows makefile for the plugin are attached. Unfortunately I do not
have access to other systems so this plugin has been tested on Windows
only.


The attached version of my plug-in has only had the copyright header
added.
I will translate this into a proper dissector rather than a plug-in as
requested, but this may take a little time as I have a lot of other
things
to do at the moment.

Me:

Convert into a normal dissector
Reorder / reformat code a bit
Added Marks name to the top of the file.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18908 f5534014-38df-0310-8fa8-9805f1628bb7