add a type field for the fid tracking structure and initialize it to UNKNOWN
when files are opened using NTCreateAndX  and if we recognize the type   set the type field to either FILE, DIR or PIPE

This is useful to know when dissecting things like security descriptors since it tells us how to dissect the specific bits of the access mask.


Only do this for NTCreateAndX for now.   It is trivial to add similar tracking to some of the older   obsolete   calls used to open fids    but no clients ever use those old calls any more.






git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@18922 f5534014-38df-0310-8fa8-9805f1628bb7