resolve the oid to a variable before we check it in the conditional so that we can add code to check that it looks sane (for better heuristics) when we start adding GSS-KRB reassembly.
we need this for some transports such as SMB/SessionSetup that will transport GSS-KRB blobs inside multiple PDUs (multiple different SMB/SessionSetups) so we can reassemble the blobs before decoding them.


this probably only happens for SMB/SessionSetup   but the design of that command is so "nice" that you can not tell whether the blob is fragmented or not or how big it is supposed to be by looking at the SMB layer itself, one needs to know the BER length field for the BER APPLICATION tag. :-(

to make things worse, the only way match multiple such fragments together one will need not just the fragments from the SessionSetup requests  but also the UID that is returned in the response to the initial request.
perverse design.


lets assume that there will almost never be multiple sessionsetups on the same tcp session in real traces    so to make things easier just ignore the UID for now when reassembling. (well reassembly is not added yet but will be)





git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19112 f5534014-38df-0310-8fa8-9805f1628bb7