From Donald White: First, the length of the header of a sub-frame may be miscalculated if if the PID field is not present, but was present in a previous sub-frame.  The calculation of the header length will use the value from the previous sub-frame.
Second, correct the typo "ength" to "length".

Third, the length of the current sub-frame was not passed as the
reported length to a sub-dissector.  When the sub-dissector calls
tvb_reported_length(), the function returns the length of the complete
frame and not the length of the sub-frame to be dissected.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19132 f5534014-38df-0310-8fa8-9805f1628bb7