Jeff Morriss: 	Fix for bug 1036
I looked at this today and found that in fact the PC stuff is pretty
hosed up in the SS7 dissectors.  For example, MTP3 *looks* OK here (DPC is
4-5-6):

    Routing label
        DPC (4-5-6) (394500)

but 394500 == 0x60504 == 6-5-4.  Something's not right.

I made a common PC dissector function for all the SS7 dissectors so as to
concentrate all this code in one place (something I've been wanting to do for a
while anyway) and fixed the reported problem as well as the above problem in
the attached patch.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19231 f5534014-38df-0310-8fa8-9805f1628bb7