we did not have a proper WERR error table for the windows errors   insterad we justkept adding the errors to the DOS error table (this merging does work ok since there is so very little overlap between he values in both ranges)
Add a new WERR error table to packet-windows-common    so that PIDL generated dissectors can use this table instead of the DOS table.


To make this table as complete and accurate as possible and to avoid having to type all the values in by hand    the table is generated from the samba doserr.h file and two small commandlines.
The commends in pakcet-windows-common.h explains how to regenerate the table from doserr.h







git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19306 f5534014-38df-0310-8fa8-9805f1628bb7