Stephen Fisher: 	packet-cisco-wireless.c is actually trying to dissect WLCCP:
I have attached a dissector I wrote from scratch for the
frames that I'm seeing.  It has #defines for the field offsets and
lengths so it should be easier to merge.  I also attached a sample
capture with one of the frames that I'm seeing.  There are more fields
in the frame I haven't yet figured out, hopefully your dissector has
those that I'm missing.

Me: - Commented in wlccp over udp as well, it works most of the time.
    - Leave the file packet-cisco-wireless.c in for the time being to
      copy over knowledge until no usable info is left in the file.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19447 f5534014-38df-0310-8fa8-9805f1628bb7