from Stephen fischer
Attached is a patch to fix bug #1163: "Dissector bug. ISO8073 COTP
protocol."  The SES dissector was incorrectly believing it had PDUs
within the COTP PDUs.  I added an additional heuristic check to see if
the length of the SES PDU is 0, then return false since it can't be zero
length.


Thanks,
  Steve






git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19733 f5534014-38df-0310-8fa8-9805f1628bb7