Julien Kerihuel:
The new version of the dissector will compile without any patches to apply to
Wireshark/NDR.pm. Jelmer has improved the patch and pidl now support enums as
return type. Regarding the nspi.cnf file, it won't be useful anymore.

Regarding the dissector, I've noticed a bug in NspiGetMatches which isn't
dumped correctly. I'm currently investigating the problem and I'll notify the
list when a fixed version is available.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19794 f5534014-38df-0310-8fa8-9805f1628bb7