If you're doing TCP reassembly by hand rather than with tcp_dissect_pdus(), pinfo->desegment_len indicates whether your dissector needs more data from TCP or not - the return value doesn't indicate that.
Fix typo.

It appears that the Id keyword is one of the case-insensitive ones in
the svn:keywords property, so if you set it to "ID" it still expands
"$Id$"; it also appears not to expand "$ID$".  We use Revision, Date,
and Author in the document to indicate the revision, and don't expand
Id, so that references to "$Id$" get left alone.

Rewrap paragraphs.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@19950 f5534014-38df-0310-8fa8-9805f1628bb7