U3 support for "recent" files.
When the "recent" file is written, any references to files on the U3 device itself are modified to have the drive replace with "$U3_DEVICE_PATH". Similarly, when the recent file is read, "$U3_DEVICE_PATH" is replaced with the current U3 device path.

This ensures that if/when the U3 device path changes when inserting it into different machines, recent captures on the U3 device itself can still be opened from the recent menu.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@20417 f5534014-38df-0310-8fa8-9805f1628bb7