From Tobias Witek:
I found a rare situation in which the BSSAP dissector seems to wrongly
assume a packet.

When a RANAP DirectTransfer message contains the GSM Supplementary
Service 'Call Confirmed' this seems to yield a message that the BSSAP
dissector recognizes as a BSSMAP BLOCK message (and from the perspective
of BSSAP, this is perfectly correct).

My patch includes code that checks this very special case.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@20520 f5534014-38df-0310-8fa8-9805f1628bb7