From Graham Bloice: In the Developers Guide, Section 9.3, Example 9.17. Decompressing data packets for dissection
The code calls tvb_set_free_cb() for the newly created next_tvb.  This
is unnecessary as the call to tvb_set_child_real_data() adds next_tvb to
the chained list of tvb, thus ensuring that next_tvb is correctly
deleted.  In fact when I had the call in, Visual Studio kept breaking
deep down in ntdll.dll, probably because of a double free every time the
main tvb was deleted.

In README.developer, para 2..2.7 The example conversation code doesn't
assign the result of conversation_new() back into the conversation variable.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@20569 f5534014-38df-0310-8fa8-9805f1628bb7