Ian Schorr <ian.schorr@gmail.com> Wed, Jan 31, 2007 at 7:24 PM To: wireshark-dev@wireshark.org
Hello,

Please consider for checkin the following new dissectors, for the FMP protocol.

FMP (File Mapping Protocol) is the network protocol basis for EMC's HighRoad (MPFS) technology.  Highroad is used to allow multiple clients to share access to NAS-shared files while allowing clients to directly access data volumes (via, for example, Fibre Channel or iSCSI).  EMC currently uses this technology in our Celerra NAS servers, and we're currently in the process of open sourcing portions of the technology. 

FMP actually consists of two ONC/RPC-based protocols - the core FMP protocol, and FMP/Notify.  The latter is used as an asynchronous callback to inform clients of status changes, such as lock revocation.

We'd like to offer these dissectors to Wireshark users for help in debugging or otherwise troubleshooting MPFS-related problems.  There are still a few minor changes that need to be made ( i.e. a handful of fields that aren't decoded) but the dissector is overall fairly complete and very usable.

Let me know if there are questions or feedback, or otherwise if other info is needed (like sample captures, which I don't want to send out to the mailing list). 

Thanks,
Ian Schorr
EMC Corporation
 



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@20679 f5534014-38df-0310-8fa8-9805f1628bb7