Fix bug #1394: Runtime error crash after clicking on HTTP packet
The capture file the user supplied had a HTTP chunked response
in it with no actual chunks other than the zero length chunk
indicating the end of the chunks.  The fix is to only create
a new_tvb and copy it over the tvb going into the
chunked_encoding_dissector() function if the chunk size is > 0.
 


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21034 f5534014-38df-0310-8fa8-9805f1628bb7