Fix bug #1375: Capture Options with many IP addresses scrolls off screen
Changed IP address label to fit into an event box which is in the ip
address hbox.  This allows the label to expand to fit the capture dialog
window, but no further.  Expanding the capture dialog window will show 
more IP addresses.  Left align the label.  Add ... in the middle of a
long list of IP addresses to show the first few and the last few with
the gtk_label_set_ellipsize() function in GTK 2.6 and up. 



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21038 f5534014-38df-0310-8fa8-9805f1628bb7