From Charles Lepple
This patch enables packet-usb.c to display all descriptors in the GET
CONFIGURATION request, regardless of whether we know how to decode
them. (All descriptors in that request share the same first two
bytes.)

It also adds the HID descriptor type (not the report descriptor,
though - that is buried in a class request, not a device request). I
am still working on actually decoding this descriptor.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21197 f5534014-38df-0310-8fa8-9805f1628bb7