From Doug Pratley:
The purpose of the patch is to provide a new output format (so it is
independent of -V): single line record per-packet with the fields chosen by the
user, with configuration options to control separator, quoting and whether a
header line is printed. It also extends some existing options behaviour (-c and
-a:filesize) so that they affect reading a file as well as writing one, so that
only the first <n> packets or bytes are read).



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21211 f5534014-38df-0310-8fa8-9805f1628bb7