Fix bug 1377:
http://bugs.wireshark.org/bugzilla/show_bug.cgi?id=1377

and complaints to the mailing list that a bogus (empty or non-ASCII contents)
dialog box was popped up when getting the list of interfaces.  The problem was
that 'get_airpcap_interface_list()' wasn't setting the returned error value
when it found that AirPcap was not loaded.  If whatever was in that variable
happened to be 1 when a non-AirPcap user requested the list of interfaces
then the bogus dialog would show up.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21380 f5534014-38df-0310-8fa8-9805f1628bb7