Use AC_HELP_STRING for the help strings for --with and --enable flags.
Add -Wpointer-arith to the GCC -W flags by default.

Make "extra-gcc-checks" and "warnings-as-errors" --enable flags rather
than --with flags - autoconf's model is that --enable is for turning
features on or off, --with is for enabling or disabling the use of
external packages (libpcap, Net-SNMP, GNU ADNS, etc.).

When testing whether the compiler is GCC, use the same style all the
time - check whether "x$GCC" equals "xyes".  (The "x" might be overkill
- if you don't quote the arguments, it avoids a missing argument to
"test"/"[", but if you do, it might not be needed.)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21492 f5534014-38df-0310-8fa8-9805f1628bb7