From Kriang Lerdsuwanakij:
This patch adds the handling of Spare Extension bytes
to UMTS Frame Protocol.  It also handles the case when
the presence of CRC in dedicated channels is not known
(i.e. when FP from a K12/K15 log is dissected).

The new functionality is placed in the new function
"dissect_spare_extension_and_crc".  The "dch_crc_present"
field inside "struct fp_info" (file packet-umts_fp.h)
is also extended to handle the case of unknown CRC presence.

Much of other changes is to update "offset" variable and
return it so that the location of Spare Extension and CRC
is available.

The patch also include a small tweak to handle Frame
Protocol information saved from K15.  Some fields appear
8 bytes later compared to K12.  The changes are in the file
packet-k12.c.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21726 f5534014-38df-0310-8fa8-9805f1628bb7