Fix for bug #1532 -
From Kenneth Golomb:

Fixes two issues:
1) interface_name is allocated iface_len+1 bytes but a buffer length of iface
len is passed in when the string is copied.  This truncates the string to 9
characters if the interface_name is 10 characters long.

2) interface_name added to tree with proto_tree_add_string_format and is
missing the initial value parameter when called.  Even if this is added the
field can not be filtered on.  Adding the field with proto_tree_add_item fixes
both the tree display and the filtering issues.



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@21730 f5534014-38df-0310-8fa8-9805f1628bb7