Add comments to various %option items to explain what they're doing. Move the %options to the beginning if they weren't already there, and put them in the same order in all files.
Add "prefix=" options to .l files that don't already have them, so we
don't have to pass a "-P" option.

Add "never-interactive" and "noyywrap" options to our lexical analyzers,
to remove extra isatty() checks and to eliminate the need for yywrap()
from the Flex library.

Get rid of %option nostdinit - that's the default.

Add .l.c: rules to Makefile.am files, replacing the rules for specific
.l files.  Have those rules all check that $(LEX) is set.

Update the address for the FSF.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@22424 f5534014-38df-0310-8fa8-9805f1628bb7