Initial Portable Apps (www.portableapps.com) packaging to produce a "Wireshark Portable" (WSP) distribution.
This allows Wireshark to be run from any USB stick, unlike the U3 packaging which must be installed on a specific U3 device.

The packaging basically builds upon the U3 distribution so new libraries/config only need to be added to the U3 makefile to update both distributions.

It still takes a "dumb" approach to WinPcap - installing it if it is not already installed and removing it (if WSP installed it) when Wireshark quits. (This is worse than the U3 packaging which only uninstalls WinPcap when the device is removed.) Really must talk to the WinPcap guys to see how we can make this less intrusive.

I am talking to John Haller (the Portable Apps guy) about fine tuning the distribution. 



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@23340 f5534014-38df-0310-8fa8-9805f1628bb7