From Florent DROUIN: Since the use of the function 'dissect_ber_tagged_type' for DialoguePortion, the file tcap.cnf must be updated to remove the decoding of the tag and length. This decoding is now done in the new function 'dissect_ber_tagged_type'.
The file tcap.cnf has been updated to take into account this change. 
But this leads to a change in tcap.asn too, for the definition of the
ExternalPDU.
I think this part of the ASN1 file is specific to Wireshark and can be
modified.

In the meantime, I did update the DEBUG part for packet_ber.c  for the function
(dissect_ber_tagged_type)

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@23442 f5534014-38df-0310-8fa8-9805f1628bb7