Bug 2116 - From  Lawrence Reisler  :
During a regular Wireshark trace of UCP packets running over ethernet, I noted
that Wireshark said the login packet was malformed, even though it looked OK
and the server responded correctly.  After looking at the UCP protocol, it
became clear that Wireshark was parsing a type 60 message which only has one
Reserverd field (RES1) and expecting it to have two Reserved fields (RES1 and
RES2) like a type 61 message.  This is because it is using the same function to
parse both messages, and does not have a conditional for the type 60 field.




git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@23903 f5534014-38df-0310-8fa8-9805f1628bb7