From Sven Meier (on -dev in 6/2007):
This is a dissector for the Parallel Redundancy Protocol (PRP) defined in chapter
6 of the IEC 62439.

PRP uses two independent networks in parallel and allows redundancy without
switchovers.

The protocol is sending Mac multicast messages with Ethertype 0x88fb. In
addition to that it adds to every Ethernet frame a 4 byte trailer before
the FCS. The trailer is detected by checking a size field and an identifier
which are part of the trailer. Therefore, if the last 4 bytes of a frame
match a correct trailer they get interpreted as a trailer, although it was
probably not a real one.

Note:

This is a post-dissector which means it gets called for every frame.  So as
to not cause a performance hit for every Wireshark user (who may not even be
looking at Ethernet frames) it currently disables itself every time Wireshark
starts up.  (There should be a better way to do this--maybe different Profiles
(as discussed on -dev recently) is the way to go.)

From me:

Put ETHERTYPE_PRP (not IANA registered) in etypes.h and packet-ethertype.c


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@24493 f5534014-38df-0310-8fa8-9805f1628bb7