From Holger Pfrommer via http://bugs.wireshark.org/bugzilla/show_bug.cgi?id=2205
This is a new dissector plugin for Hilscher analyzer frames.
These frames are generated by Hilscher analyzer products and are identified via
their unique source MAC address (this is a reserved MAC from Hilscher-range and
will never be used by another network device). Most likely these frames are
only generated on a virtual network interface or the generating device is
attached directly via patch cable to a real network interface, but not routed
through a network. The Ethernet-header (destination MAC, source MAC and
Length/Type) is not displayed in the protocol tree for these frames as this is
overhead-information which has no practical use in this case.

Note:

This is a heuristic Ethernet dissector which means it gets called for every
Ethernet frame.  So as to not cause a performance hit for most Wireshark users
it has a preference which, by default, disables the dissector.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@24495 f5534014-38df-0310-8fa8-9805f1628bb7