From Gerhard Olsson: Add "Assume COTP" prefererence instead of hard coding
If there is no current dissector registered for X25 and there is no hints in
the call packets, X25 tries to look at user data, then try heuristics. See end
of dissect_x25_common

This means that if the call data happens to start with 0x45, IP is assumed as
the data format even if the packet cannot be handled as IP (for instance too
short).

It is better to try the heuristics first as they should have more complete tests.

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@24655 f5534014-38df-0310-8fa8-9805f1628bb7