Allow the folders in the About Wireshark/Folders list to be double clicked on to open the appropriate folder in explorer.
The MIB/PIB paths have been split up to allow each one to be clicked on separately, but as a result this will lose the order if the whole list is reordered.

XXX: I have been unable to test this on MacOS or Linux. Stig? :-)
I suppose it should work, but we may need to make it WIN32 specific?



git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@24995 f5534014-38df-0310-8fa8-9805f1628bb7