Fix Dissector bug, protocol SNMP: proto.c:932: failed assertion From LEGO: DISSECTOR_ASSERT_NOT_REACHED() is a Bug regardless, it is triggered because we are proto_add_item()ing an FT_UINT32 of 5 bytes in length.
IF-MIB:ifInOctets is a Counter32 so the value should (and is) application
encoded using tag 41.

the value's value (!) is  H'00d49e69fa (D'3567151610).

the assertion is triggered because there's an error in oids.c:59 
where states the ft_type as being an FT_UINT32 with a length of 1 to 4 bytes,
It should be FT_UINT64 and the length from 1 to 5 bytes considering that
integers bigger than 2^31 will be ber encoded in 5 bytes.

http://bugs.wireshark.org/bugzilla/show_bug.cgi?id=2468

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@25148 f5534014-38df-0310-8fa8-9805f1628bb7