g_strcmp0() only shows up on the Intarweb in documentation for some testing version of GLib; it doesn't appear to exist in any current versions.  There's no need to "gracefully" handle NULL (whatever "gracefully" means in this context - NULL compares less than any real string?), as we already checked whether mac_name is null, and we're comparing it against a string constant; just use strcmp().
In ssh_gslist_compare_strings(), check for null pointers for now.

strcmp() (and strcmp-alikes) don't return a boolean, they return a value
such that comparing strings with a particular operator is done by
comparing the result of strcmp() with 0 using that operator; do that, to
make it clearer that the strings are being compared for equality.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@25402 f5534014-38df-0310-8fa8-9805f1628bb7