Remove a lot of vestigal GTK-Wimp stuff. It's been included with GTK+ since version 2.8, when its name was changed to the "GTK MS Windows-Engine."
Simplify the Windows installer by removing the GTK-Wimp option (which
means the "Wireshark" item can be an entry instead of a group). Move
Rawshark to the "Tools" group.

(If anyone wants to disable the MS Windows Engine they can still do so
after installation by disabling or removing libwimp.dll.)


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@25611 f5534014-38df-0310-8fa8-9805f1628bb7