From Alexey Neyman : This patch implements a function for dissecting bitfields with better control over the resulting representation than the existing proto_tree_add_bitmask() routine. This function will be used by reworked IPMI/ATCA dissector (bug 2048).
The function is described in README.developer. In short, the differences are as
follows:

- The new function does not require a hf_XXX field for the whole bitmask. When
the bitmask includes several unrelated fields, such hf_XXX field does not make
sense.

- The new function allows better control over the way the sub-item descriptions
are added to the top-level item. For example, proto_tree_add_bitmask() function
does not add non-enumerated integers, does not use true_false_string to display
boolean.

- The new function allows to specify "fallback" text for the top-level item
which is used if no items were added to the top-level item.

git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@25920 f5534014-38df-0310-8fa8-9805f1628bb7